Fix some issues with contextIsolation
