--- packages/app/src/static/preload/whatsappPreload.js
@@ -1,15 +1,0 @@
-let snoozed = false;
-const originalPlay = Audio.prototype.play;
-
-const shouldBypassSnooze = (audio) => audio.src.startsWith('blob:http');
-
-Audio.prototype.play = function() {
-  if (!snoozed || shouldBypassSnooze(this)) {
-    return originalPlay.call(this);
-  }
-  return Promise.resolve();
-};
-
-window.bx.notificationCenter.addSnoozeDurationInMsChangeListener((_, duration) => {
-  snoozed = Boolean(duration);
-});
