+++ packages/appstore/index.html
@@ -18,6 +18,9 @@
         </style>
     </head>
     <body>
+        <script>
+          var global = global || window; //vk: https://github.com/webpack/webpack/issues/10035
+        </script>
         <div id="root"></div>
     </body>
 </html>
